+++ laktory/models/datasinks/basedatasink.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import Union
 from typing import Any
 from typing import Literal
-from pydantic import Field
+from pydantic import model_validator
 from laktory._logger import get_logger
 from laktory.models.basemodel import BaseModel
 from laktory.spark import is_spark_dataframe
@@ -16,6 +16,440 @@ from laktory.polars import PolarsLazyFrame
 from laktory.types import AnyDataFrame
 
 logger = get_logger(__name__)
+
+
+class DataSinkMergeCDCOptions(BaseModel):
+    """
+    Options for merging a change data capture (CDC).
+
+    They are also used to build the target using `apply_changes` method when
+    using Databricks DLT.
+
+    Attributes
+    ----------
+    delete_where:
+        Specifies when a CDC event should be treated as a DELETE rather than
+        an upsert.
+    end_at_column_name:
+        When using SCD type 2, name of the column storing the end time (or
+        sequencing index) during which a row is active. This attribute is not
+        used when using Databricks DLT which does not allow column rename.
+    include_columns:
+        A subset of columns to include in the target table. Use
+        `include_columns` to specify the complete list of columns to include.
+    exclude_columns:
+        A subset of columns to exclude in the target table.
+    ignore_null_updates:
+        Allow ingesting updates containing a subset of the target columns.
+        When a CDC event matches an existing row and ignore_null_updates is
+        `True`, columns with a null will retain their existing values in the
+        target. This also applies to nested columns with a value of null. When
+        ignore_null_updates is `False`, existing values will be overwritten
+        with null values.
+    order_by:
+        The column name specifying the logical order of CDC events in the
+        source data. Used to handle change events that arrive out of order.
+    primary_keys:
+        The column or combination of columns that uniquely identify a row in
+        the source data. This is used to identify which CDC events apply to
+        specific records in the target table.
+    scd_type:
+        Whether to store records as SCD type 1 or SCD type 2.
+    start_at_column_name:
+        When using SCD type 2, name of the column storing the start time (or
+        sequencing index) during which a row is active. This attribute is not
+        used when using Databricks DLT which does not allow column rename.
+
+    References
+    ----------
+    https://iterationinsights.com/article/how-to-implement-slowly-changing-dimensions-scd-type-2-using-delta-table/
+    https://docs.databricks.com/en/delta-live-tables/python-ref.html#change-data-capture-with-python-in-delta-live-tables
+    """
+
+    # apply_as_truncates: Union[str, None] = None
+    delete_where: str = None
+    end_at_column_name: str = "__end_at"
+    exclude_columns: list[str] = None
+    ignore_null_updates: bool = False
+    include_columns: list[str] = None
+    order_by: str = None
+    primary_keys: list[str]
+    scd_type: Literal[1, 2] = 1
+    start_at_column_name: str = "__start_at"
+    # track_history_columns: Union[list[str], None] = None
+    # track_history_except_columns: Union[list[str], None] = None
+    _parent: Any = None
+    _source_schema: Any = None
+    _source_columns: list[str] = None
+
+    @model_validator(mode="after")
+    def validate_scd_type(self) -> Any:
+        if self.scd_type == 2:
+            if self.order_by is None:
+                raise ValueError(
+                    "SCD Type 2 merge requires specific of `order_by` attribute."
+                )
+
+        return self
+
+    @model_validator(mode="after")
+    def selected_columns(self) -> Any:
+        if self.include_columns and self.exclude_columns:
+            raise ValueError(
+                "`include_columns` and `exclude_columns` attributes are mutually exclusive."
+            )
+
+        return self
+
+    # ----------------------------------------------------------------------- #
+    # Sink                                                                    #
+    # ----------------------------------------------------------------------- #
+
+    @property
+    def sink(self):
+        return self._parent
+
+    @property
+    def target_name(self):
+        if self._parent and "TableDataSink" in str(type(self._parent)):
+            return self._parent.full_name
+        return None
+
+    @property
+    def target_path(self):
+        if self._parent and "FileDataSink" in str(type(self._parent)):
+            return self._parent.path
+        return None
+
+    @property
+    def target_id(self):
+        return self.target_name or self.target_path
+
+    # ----------------------------------------------------------------------- #
+    # CDC Columns                                                             #
+    # ----------------------------------------------------------------------- #
+
+    @property
+    def index(self):
+        return self.order_by
+
+    @property
+    def start_at(self):
+        return self.start_at_column_name
+
+    @property
+    def end_at(self):
+        return self.end_at_column_name
+
+    @property
+    def index_fist(self):
+        return self.index + "_first"
+
+    @property
+    def hash_keys(self):
+        return "__hash_keys"
+
+    @property
+    def hash_cols(self):
+        return "__hash_cols"
+
+    @property
+    def source_columns(self):
+        return [f.name for f in self._source_schema]
+
+    @property
+    def update_columns(self):
+        if self.include_columns:
+            columns = [c for c in self.include_columns]
+        else:
+            columns = [c for c in self.source_columns if c not in self.primary_keys]
+            if self.exclude_columns:
+                columns = [c for c in columns if c not in self.exclude_columns]
+        return columns
+
+    @property
+    def extra_columns(self):
+        cols = [self.hash_keys]
+        if self.scd_type == 2:
+            cols += [self.start_at, self.end_at, self.hash_cols]
+        return cols
+
+    @property
+    def write_columns(self):
+        return self.primary_keys + self.update_columns + self.extra_columns
+
+    @property
+    def index_type(self):
+        return [
+            field.dataType
+            for field in self._source_schema.fields
+            if field.name == self.index
+        ][0]
+
+    @property
+    def source_delete_where(self):
+        return self._add_alias(self.delete_where)
+
+    # ----------------------------------------------------------------------- #
+    # Methods                                                                 #
+    # ----------------------------------------------------------------------- #
+
+    @staticmethod
+    def _add_alias(expr, prefix="source"):
+
+        operators = ["=", ">", "<", "!", "*", "+", "-", "/", ","]
+
+        new_expr = expr
+        for o in operators:
+            new_expr = new_expr.replace(o, " ")
+
+        candidates = [c for c in new_expr.split(" ") if c != ""]
+
+        # Define other exclusions for SQL keywords or constants
+        exclusions = [
+            "TRUE",
+            "FALSE",
+            "CASE",
+            "WHEN",
+            "THEN",
+            "SELECT",
+            "FROM",
+            "WHERE",
+            "AND",
+            "OR",
+            "AS",
+            "END",
+            "ELSE",
+            "NOT",
+            "NULL",
+        ]
+
+        # Filter out identifiers that are constants or keywords
+        column_names = []
+        for c in candidates:
+            if c.upper() in exclusions:
+                continue
+            if c.startswith("'"):
+                continue
+            try:
+                float(c)
+                continue
+            except ValueError:
+                pass
+
+            column_names += [c]
+
+        new_expr = expr
+        for c in column_names:
+            if f"{prefix}." not in c:
+                new_expr = new_expr.replace(c, f"{prefix}.{c}")
+
+        return new_expr
+
+    def _init_target(self, source):
+
+        import pyspark.sql.types as T
+
+        spark = source.sparkSession
+        logger.info(f"Merge target not found. Creating empty table at {self.target_id}")
+        schema = source.select(self.primary_keys + self.update_columns).schema
+        schema.add(T.StructField(self.hash_keys, T.StringType(), True))
+        if self.scd_type == 2:
+            schema.add(T.StructField(self.hash_cols, T.StringType(), True))
+            schema.add(T.StructField(self.start_at, self.index_type, True))
+            schema.add(T.StructField(self.end_at, self.index_type, True))
+
+        df = spark.createDataFrame(data=[], schema=schema)
+
+        writer = df.write.format("DELTA").mode("OVERWRITE")
+        if self.target_path:
+            writer.save(self.target_path)
+        else:
+            writer.saveAsTable(self.target_name)
+
+    def _execute(self, source: SparkDataFrame):
+
+        from delta.tables import DeltaTable
+        from pyspark.sql import Window
+        import pyspark.sql.functions as F
+
+        spark = source.sparkSession
+
+        logger.info(
+            f"Executing merge on {self.target_id} with primary keys {self.primary_keys} and scd type {self.scd_type}"
+        )
+
+        # Add internal columns
+        source = source.withColumn(
+            self.hash_keys, F.lit(F.sha2(F.concat_ws("~", *self.primary_keys), 256))
+        )
+        if self.scd_type == 2:
+            source = (
+                source.withColumn(self.start_at, F.col(self.index))
+                .withColumn(self.end_at, F.lit(None).cast(self.index_type))
+                .withColumn(
+                    self.hash_cols,
+                    F.lit(F.sha2(F.concat_ws("~", *self.update_columns), 256)),
+                )
+            )
+
+        # Process History
+        if self.index:
+            w = Window.partitionBy(*self.primary_keys).orderBy(F.desc(self.index))
+            source = source.withColumn("_row_number", F.row_number().over(w))
+            if self.scd_type == 1:
+                # Drop Duplicates
+                source = source.filter(F.col("_row_number") == 1)
+            elif self.scd_type == 2:
+                # Assign previous index to ends_at
+                w2 = Window.partitionBy(*self.primary_keys)
+                source = source.withColumn(self.end_at, F.lag(self.index, 1).over(w))
+                source = source.withColumn(self.index_fist, F.min(self.index).over(w2))
+            source = source.drop("_row_number")
+
+        # Read target
+        if self.target_path:
+            table_target = DeltaTable.forPath(spark, self.target_path)
+        else:
+            table_target = DeltaTable.forName(spark, self.target_name)
+
+        if self.scd_type == 1:
+
+            # Define merge
+            merge = table_target.alias("target").merge(
+                source.alias("source"),
+                condition=f"source.{self.hash_keys} = target.{self.hash_keys}",
+            )
+
+            # Update
+            _set = {f"target.{c}": f"source.{c}" for c in self.update_columns}
+            if self.ignore_null_updates:
+                _set = {
+                    f"target.{c}": F.coalesce(
+                        F.col(f"source.{c}"), F.col(f"target.{c}")
+                    ).alias(c)
+                    for c in self.update_columns
+                }
+
+            condition = None
+            if self.delete_where:
+                condition = ~F.expr(self.source_delete_where)
+            merge = merge.whenMatchedUpdate(set=_set, condition=condition)
+
+            # Insert
+            merge = merge.whenNotMatchedInsert(
+                values={f"target.{c}": f"source.{c}" for c in self.write_columns}
+            )
+
+            # Delete
+            if self.delete_where:
+                merge = merge.whenMatchedDelete(
+                    condition=F.expr(self.source_delete_where)
+                )
+
+            merge.execute()
+
+        elif self.scd_type == 2:
+
+            # Only select rows that have been updated
+            if self.target_path:
+                target = spark.read.format("DELTA").load(self.target_path)
+            else:
+                target = spark.read.table(self.target_name)
+            upsert_or_delete = source.withColumn(
+                "__to_delete", F.expr(self.delete_where)
+            ).join(
+                other=target.withColumn("__to_delete", F.lit(False)),
+                on=[self.hash_cols, self.hash_keys, "__to_delete"],
+                how="leftanti",
+            )
+
+            # Merge
+            condition = F.expr(f"source.{self.hash_keys} = target.{self.hash_keys}")
+            condition = condition & F.expr(f"target.{self.end_at} IS NULL")
+            merge = table_target.alias("target").merge(
+                upsert_or_delete.filter(F.col(self.end_at).isNull()).alias("source"),
+                condition=condition,
+            )
+
+            # Expire the current record
+            _set = {f"target.{self.end_at}": f"source.{self.index_fist}"}
+            merge = merge.whenMatchedUpdate(set=_set)
+            #
+            # # TODO: Review if required
+            # if not self.delete_where:
+            #     _set = {f"target.{self.end_at}": f"source.{self.order_by}"}
+            #     merge = merge.whenMatchedUpdate(set=_set)
+            # else:
+            #     where = F.expr(self.source_delete_where)
+            #     # deleting
+            #     # _set = {f"target.{self.end_at}": "NULL"}
+            #     _set = {f"target.{self.end_at}": f"source.{self.order_by}"}
+            #     merge = merge.whenMatchedUpdate(set=_set, condition=where)
+            #
+            #     # updating
+            #     _set = {f"target.{self.end_at}": f"source.{self.order_by}"}
+            #     merge = merge.whenMatchedUpdate(set=_set, condition=~where)
+
+            merge.execute()
+
+            # Append rows
+            upsert = upsert_or_delete
+            if self.delete_where:
+                upsert = upsert.filter(~F.expr(self.delete_where))
+            writer = (
+                upsert.select(self.write_columns)
+                .write.mode("APPEND")
+                .format("DELTA")
+            )
+            if self.target_path:
+                writer.save(self.target_path)
+            else:
+                writer.saveAsTable(self.target_name)
+
+        else:
+            raise ValueError(f"SCD Type {self.scd_type} is not supported.")
+
+    def execute(self, source: SparkDataFrame):
+
+        from delta.tables import DeltaTable
+
+        self._source_schema = source.schema
+        spark = source.sparkSession
+
+        if self.target_path:
+            if not DeltaTable.isDeltaTable(spark, self.target_path):
+                self._init_target(source)
+        else:
+            try:
+                spark.catalog.getTable(self.target_name)
+            except Exception as e:
+                self._init_target(source)
+
+        if source.isStreaming:
+
+            if self.sink is None:
+                raise ValueError(f"Sink value required to fetch checkpoint location.")
+
+            if self.sink and self.sink._checkpoint_location is None:
+                raise ValueError(
+                    f"Checkpoint location not specified for sink '{self.sink}'"
+                )
+
+            query = (
+                source.writeStream.foreachBatch(
+                    lambda batch_df, batch_id: self._execute(source=batch_df)
+                )
+                .trigger(availableNow=True)
+                .options(
+                    checkpointLocation=self.sink._checkpoint_location,
+                )
+                .start()
+            )
+            query.awaitTermination()
+
+        else:
+            self._execute(source=source)
 
 
 class BaseDataSink(BaseModel):
@@ -29,6 +463,9 @@ class BaseDataSink(BaseModel):
         moving from stream to batch. Don't apply for quarantine sinks.
     is_quarantine:
         Sink used to store quarantined results from node expectations.
+    merge_cdc_options:
+        Merge options to handle input DataFrames that are Change Data Capture
+        (CDC). Only used when `merge` mode is selected.
     mode:
         Write mode.
         - overwrite: Overwrite existing data
@@ -36,6 +473,8 @@ class BaseDataSink(BaseModel):
         - error: Throw and exception if data already exists
         - ignore: Silently ignore this operation if data already exists
         - complete: Overwrite for streaming dataframes
+        - merge: Append, update and optionally delete records. Requires
+        cdc specification.
     write_options:
         Other options passed to `spark.write.options`
     """
@@ -43,11 +482,27 @@ class BaseDataSink(BaseModel):
     is_quarantine: bool = False
     is_primary: bool = True
     checkpoint_location: str = None
+    merge_cdc_options: DataSinkMergeCDCOptions = None  # TODO: Review parameter name
     mode: Union[
-        Literal["OVERWRITE", "APPEND", "IGNORE", "ERROR", "COMPLETE", "UPDATE"], None
+        Literal[
+            "OVERWRITE", "APPEND", "IGNORE", "ERROR", "COMPLETE", "UPDATE", "MERGE"
+        ],
+        None,
     ] = None
     write_options: dict[str, str] = {}
     _parent: "PipelineNode" = None
+
+    @model_validator(mode="after")
+    def merge_has_options(self) -> Any:
+        if self.mode == "MERGE":
+            if self.merge_cdc_options is None:
+                raise ValueError(
+                    "If 'MERGE' `mode` is selected, `merge_cdc_options` must be specified."
+                )
+            else:
+                self.merge_cdc_options._parent = self
+
+        return self
 
     # ----------------------------------------------------------------------- #
     # Properties                                                              #
@@ -77,6 +532,39 @@ class BaseDataSink(BaseModel):
                     )
 
         return None
+
+    # ----------------------------------------------------------------------- #
+    # CDC                                                                     #
+    # ----------------------------------------------------------------------- #
+
+    @property
+    def is_cdc(self) -> bool:
+        return self.merge_cdc_options is not None
+
+    @property
+    def dlt_apply_changes_kwargs(self) -> dict[str, str]:
+        """Keyword arguments for dlt.apply_changes function"""
+
+        from laktory.models.datasinks.tabledatasink import TableDataSink
+
+        if not isinstance(self, TableDataSink):
+            raise ValueError("DLT only supports `TableDataSink` class")
+
+        cdc = self.merge_cdc_options
+        return {
+            "apply_as_deletes": cdc.delete_where,
+            # "apply_as_truncates": ,  # NOT SUPPORTED
+            "column_list": cdc.include_columns,
+            "except_column_list": cdc.exclude_columns,
+            "ignore_null_updates": cdc.ignore_null_updates,
+            "keys": cdc.primary_keys,
+            "sequence_by": cdc.order_by,
+            # "source": self.source.table_name,  # TO SET EXTERNALLY
+            "stored_as_scd_type": cdc.scd_type,
+            "target": self.table_name,
+            # "track_history_column_list": cdc.track_history_columns,  # NOT SUPPORTED
+            # "track_history_except_column_list": cdc.track_history_except_columns,  # NOT SUPPORTED
+        }
 
     # ----------------------------------------------------------------------- #
     # Writers                                                                 #